support-code: put code at correct place
