+++ cucumber-support-code/javascript/test/CucumberSupportCodeTest.ts
@@ -1,74 +1,8 @@
 import assert from 'assert'
 import { messages } from 'cucumber-messages'
-import ISupportCodeExecutor from '../src/ISupportCodeExecutor'
+
+import CucumberSupportCode from '../src/CucumberSupportCode'
 import SupportCodeExecutor from '../src/SupportCodeExecutor'
-import uuidv4 from 'uuid/v4'
-import parseTagExpression from 'cucumber-tag-expressions'
-import { isMainThread } from 'worker_threads'
-import { isContext } from 'vm'
-
-class Hook {
-  public readonly id: string = uuidv4()
-
-  constructor(
-    private readonly tagExpression: string,
-    private readonly executor: ISupportCodeExecutor
-  ) {}
-
-  public match(tags: string[]): boolean {
-    const expression = parseTagExpression(this.tagExpression)
-    return expression.evaluate(tags)
-  }
-
-  public execute(): any {
-    return this.executor.execute()
-  }
-}
-
-class CucumberSupportCode {
-  private beforeHooks: Hook[] = []
-  private hookById: { [key: string]: Hook; } = {}
-
-  public registerBeforeHook(
-    tagExpression: string,
-    executor: ISupportCodeExecutor
-  ): messages.ITestCaseHookDefinitionConfig {
-    const hook = new Hook(tagExpression, executor)
-    this.beforeHooks.push(hook)
-    this.hookById[hook.id] = hook
-
-    return new messages.TestCaseHookDefinitionConfig({
-      id: hook.id,
-      tagExpression: tagExpression
-    })
-  }
-
-  public findBeforeHooks(tags: string[]): string[] {
-    return this.beforeHooks
-      .map(hook => hook.match(tags) ? hook.id : undefined)
-      .filter(hookId => hookId !== undefined)
-  }
-
-  public executeHook(hookId: string): messages.ITestResult {
-    const hook = this.hookById[hookId]
-    if (hook === undefined) {
-      throw new Error("Hook not found")
-    }
-
-    try {
-      hook.execute()
-
-      return new messages.TestResult({
-        status: messages.TestResult.Status.PASSED,
-      })
-    } catch (error) {
-      return new messages.TestResult({
-        status: messages.TestResult.Status.FAILED,
-        message: error.stack
-      })
-    }
-  }
-}
 
 describe('CucumberSupportCode', () => {
   describe('#registerBeforeHook', () => {
