+++ fake-cucumber/javascript/.internal-dependencies
@@ -3,3 +3,4 @@ cucumber-expressions
 c21e
 gherkin
 tag-expressions
+cucumber-support-code
